+++ HelperFunctions.py
@@ -1,8 +1,0 @@
-import json
-
-
-def load_file_json(file_name):
-	with open(file_name, 'r') as _file:
-		content = _file.read()
-		content_dict = json.loads(content)
-		return content_dict