+++ changelog.md
@@ -9,3 +9,10 @@
         - inline requests supported
         - inline error showing
         - about with emojis
+
++ Version 0.18.1.2
+
+        - ajout de l'API coinmarketcap
+            -création de la liste des monnaies (nom/symbole)
+            -enregistrement de cette liste sur le disque (cache)
+        - ajout d'une fonction helper (save_json)